completely changed the debug display output again using tview
This is still hacked in. For sure I have no real idea what I am doing
with tview, and I have no idea why calling application.Stop() will
segfault. I am relying on Ctrl-C for now to get out of the app (which
works).

BUT: we finally see colors, hooray. The display still uses the utf-8
blocks to express the intensity, and scales the colors up to max
intensity for the devug output so we can see them better (even if they
are low in intensity on the real hardware)

On top of that another magic scaling happens to account for the
dramatic color shift of the real hardware:
RGB 80/15/3 is giving a warmish white with my led strips.
So the components are way off in their color reproduction.
The end result is a somewhat pleasing looking color rendition on the
terminal, but it is somehow tied to the very hardware I am using.
This could change if I would use a correction factor for the display
of the LEDs, and assume 255/255/255 for full white in the
calculations. But hey, why bother