refactor: Move Trigger to util to break cycle
Moved the Trigger struct from the platform package to the new util
package. This resolves a cyclic dependency between the platform and
producer packages.

The LedProducer interface and all its implementations were updated to
accept a '*util.Trigger' instead of a 'time.Time'. The main application
logic in goleds.go and all tests were updated accordingly.