config file creation from default, incl. handling missing parameters
if the config file is missing, it is created by copying
config.yml.orig. If it exists, config.yml.orig is still read into the
config structure first, followed by reading the config file, thereby
supplying missing (or new) parameters in the config file from the
default values.

The config file is not automatically updated by these sections. Only
if saved by the system (by changes through the web service) the
missing sections are filled in.